[수정] 스크롤 카드 플립 애니메이션 커스텀훅을로 분리
